[BE/FIX/#12] ENTITY 수정 확인

See merge request s12-webmobile1-sub1/S12P11A506!25

+++ Backend/src/main/java/com/hexa/muinus/store/domain/store/repository/StoreRepository.java
@@ -9,12 +9,13 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 public interface StoreRepository extends JpaRepository<Store, Integer> {
 
-    Optional<Store> findByLocationXAndLocationY(double locationX, double locationY);
+    Optional<Store> findByLocationXAndLocationY(BigDecimal locationX, BigDecimal locationY);
     Optional<Store> findByUserAndStoreNo(Users user, Integer storeNo);
 
     @Query(value = """
